requirements.txt
attrs==19.3.0
aws-cdk.assets==1.19.0
aws-cdk.aws-apigateway==1.19.0
aws-cdk.aws-applicationautoscaling==1.19.0
aws-cdk.aws-autoscaling==1.19.0
aws-cdk.aws-autoscaling-common==1.19.0
aws-cdk.aws-autoscaling-hooktargets==1.19.0
aws-cdk.aws-certificatemanager==1.19.0
aws-cdk.aws-cloudformation==1.19.0
aws-cdk.aws-cloudfront==1.19.0
aws-cdk.aws-cloudwatch==1.19.0
aws-cdk.aws-ec2==1.19.0
aws-cdk.aws-ecr==1.19.0
aws-cdk.aws-ecr-assets==1.19.0
aws-cdk.aws-ecs==1.19.0
aws-cdk.aws-elasticloadbalancing==1.19.0
aws-cdk.aws-elasticloadbalancingv2==1.19.0
aws-cdk.aws-events==1.19.0
aws-cdk.aws-iam==1.19.0
aws-cdk.aws-kms==1.19.0
aws-cdk.aws-lambda==1.19.0
aws-cdk.aws-logs==1.19.0
aws-cdk.aws-rds==1.19.0
aws-cdk.aws-route53==1.19.0
aws-cdk.aws-route53-targets==1.19.0
aws-cdk.aws-s3==1.19.0
aws-cdk.aws-s3-assets==1.19.0
aws-cdk.aws-sam==1.19.0
aws-cdk.aws-secretsmanager==1.19.0
aws-cdk.aws-servicediscovery==1.19.0
aws-cdk.aws-sns==1.19.0
aws-cdk.aws-sns-subscriptions==1.19.0
aws-cdk.aws-sqs==1.19.0
aws-cdk.aws-ssm==1.19.0
aws-cdk.core==1.19.0
aws-cdk.cx-api==1.19.0
aws-cdk.region-info==1.19.0
cattrs==0.9.0
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsii==0.20.11
MarkupSafe==1.1.1
publication==0.0.3
python-dateutil==2.8.1
six==1.13.0
typing-extensions==3.7.4.1
Werkzeug==0.16.0
